chore: create generator API section in readme
input from test-cases with field category
will inject new documentation in section api
